move class UntoastedDatum to k2_util.h and address PR comments
